Move keystore to `SDK` utils
